Add files via upload
Passenger Category
A pie chat was used to group certain categories of passengers. Those passengers that appear 6 and above on a trip was taken out of the analyses.
The result shows that most trips had one passenger in then followed by two and five passengers the least is four.
The Cost Differences png Graph  shows the pickup and drop-off period of passengers from the analysis there is a difference between pickup and drop off. The both analysis of pickup latitude and drop-off longitude shows must trip were short and there is a low density of pickup. The Analysis shows the longest trip of 41.6 next to trip 41.4. 
Recurrence-The analysis shows that there is mostly one passenger on a trip. While a few times you have two passengers on a trip.
Conclusion The further the distance the higher in the cost of a trip.